format code

+++ src/if_else.sol
@@ -2,17 +2,12 @@
 pragma solidity ^0.8.17;
 
 contract IfElse {
-    function foo(uint x) public pure returns (uint res) {
+    function foo(uint256 x) public pure returns (uint256 res) {
         assembly {
-            if lt(x, 10){
-                res := 0
-            }
-            if lt(x, 20){
-                res := 1
-            }
+            if lt(x, 10) { res := 0 }
+            if lt(x, 20) { res := 1 }
             res := 2
         }
         return res;
     }
-
 }
