+++ src/simple_storage.sol
@@ -4,10 +4,10 @@ pragma solidity ^0.8.17;
 // https://solidity-by-example.org/state-variables/
 contract SimpleStorage {
     // State variable to store a number
-    uint public num;
+    uint256 public num;
 
     // You need to send a transaction to write to a state variable.
-    function set(uint _num) public {
+    function set(uint256 _num) public {
         num = _num;
         assembly {
             sstore(0, _num)
@@ -15,7 +15,7 @@ contract SimpleStorage {
     }
 
     // You can read from a state variable without sending a transaction.
-    function get() public view returns (uint num_) {
+    function get() public view returns (uint256 num_) {
         assembly {
             num_ := sload(0)
         }
